Fold testing against different database engines into single server
So that one server can be used to test against different database
engines and doesn't need to be restarted to test a different engine.
All engines are connected at startup and the engine to run against is
selected by the client by means of a URL path parameter.

Remove unnecessary combinations of docker based benchmark servers, it
is now enough to have one per web engine (Resin, Jetty, Undertow).
Also remove links definitions as they are not needed in compose v2+
(services in a stack are by default addressable by their name).

Ref #97